refactor: move old tests away, add new test harness
