+++ bash-tests/Containerfile
@@ -1,8 +1,8 @@
 FROM debian:bullseye-20211220-slim
 
 # install latest nvim
-RUN apt-get update && \
-    apt-get install -y curl jq uuid-runtime && \
+RUN apt update && \
+    apt install -y exa curl jq wget && \
     curl -sL https://api.github.com/repos/neovim/neovim/releases/latest | \
     jq -r '.assets[].browser_download_url' | \
     grep nvim-linux64.tar.gz | \
@@ -12,13 +12,21 @@ RUN apt-get update && \
     useradd -ms /bin/bash user
 
 RUN ln -s /nvim-linux64/bin/nvim /bin/nvim
+COPY ./test/assert.sh /usr/bin/assert
+COPY ./test/fail.sh /usr/bin/fail
 
 # Make sure you build with hotpot's root dir as the build context!
 COPY ./ /home/user/hotpot
 
+# copy test files
+COPY ./test /home/user/test
+RUN ln -s /home/user/test /home/user/test/lua && \
+    mkdir -p /home/user/config/fnl
+
 RUN chown -R user:user /home/user
 USER user
-WORKDIR /home/user/hotpot
+WORKDIR /home/user
 
+# we want the container to keep running and we'll exec commands in it.
+# this is so we can run lua and sh in stages, on the same fs
 ENTRYPOINT ["/bin/bash"]
-CMD ["test/test.sh"]
