+++ bash-tests/api-make/test.sh
