+++ bash-tests/api/cmd_fnl.fnl
