+++ bash-tests/api/test-compile.fnl
