+++ bash-tests/api/test-eval.fnl
