+++ bash-tests/cache_invalidation/my_macro.fnl
