+++ bash-tests/cache_invalidation/test.lua
