+++ bash-tests/cache_invalidation/test.sh
