+++ bash-tests/fail.sh
