+++ bash-tests/harness.sh
