+++ bash-tests/init.lua
