+++ bash-tests/nvim.sh
