+++ bash-tests/plugins/force_compile.fnl
