+++ bash-tests/plugins/plugin.fnl
