+++ bash-tests/plugins/test.lua
