+++ bash-tests/plugins/test.sh
