+++ bash-tests/require_hotpot/test.lua
