+++ bash-tests/require_hotpot/test.sh
