+++ test-init.fnl
@@ -1,0 +1,5 @@
+(vim.opt.runtimepath:prepend "/hotpot")
+(require :hotpot)
+(print :hi)
+(print (vim.inspect (require :hotpot)))
+(require :asrtasrt)
