+++ test-init.lua
@@ -1,0 +1,5 @@
+do end (vim.opt.runtimepath):prepend("/hotpot")
+require("hotpot")
+print("hi")
+print(vim.inspect(require("hotpot")))
+return require("asrtasrt")
