+++ test/require-a-fnl-file.lua
@@ -1,4 +1,4 @@
-package.preload["new-tests.utils"] = package.preload["new-tests.utils"] or function(...)
+package.preload["test.utils"] = package.preload["test.utils"] or function(...)
   local function read_file(path)
     return table.concat(vim.fn.readfile(path), "\\n")
   end
@@ -31,11 +31,11 @@ package.preload["new-tests.utils"] = package.preload["new-tests.utils"] or funct
     print("\n")
     return os.exit(results.fails)
   end
-  do end (vim.opt.runtimepath):prepend("/hotpot")
+  do end (vim.opt.runtimepath):prepend(vim.loop.cwd())
   require("hotpot")
   return {["write-file"] = write_file, ["read-file"] = read_file, OK = OK, FAIL = FAIL, exit = exit, NVIM_APPNAME = vim.env.NVIM_APPNAME}
 end
-local _local_1_ = require("new-tests.utils")
+local _local_1_ = require("test.utils")
 local FAIL = _local_1_["FAIL"]
 local NVIM_APPNAME = _local_1_["NVIM_APPNAME"]
 local OK = _local_1_["OK"]
@@ -66,10 +66,12 @@ local function test_path(modname, path)
           local function _12_(...)
             local _11_ = read_file(lua_path)
             if (_11_ == "return {works = true}") then
-              return OK(string.format("Outputs correct lua code"))
+              OK(string.format("Outputs correct lua code"))
+              return true
             elseif true then
               local __1_auto = _11_
-              return FAIL(string.format("Outputs correct lua code"))
+              FAIL(string.format("Outputs correct lua code"))
+              return false
             else
               return nil
             end
@@ -85,10 +87,12 @@ local function test_path(modname, path)
       local function _16_(...)
         local _15_ = vim.loop.fs_access(lua_path, "R")
         if (_15_ == true) then
-          return OK(string.format("Creates a lua file at %s", lua_path))
+          OK(string.format("Creates a lua file at %s", lua_path))
+          return true
         elseif true then
           local __1_auto = _15_
-          return FAIL(string.format("Creates a lua file at %s", lua_path))
+          FAIL(string.format("Creates a lua file at %s", lua_path))
+          return false
         else
           return nil
         end
@@ -104,10 +108,12 @@ local function test_path(modname, path)
   local function _21_()
     local _19_, _20_ = pcall(require, modname)
     if ((_19_ == true) and ((_G.type(_20_) == "table") and ((_20_).works == true))) then
-      return OK(string.format("Can require module %s %s", modname, fnl_path))
+      OK(string.format("Can require module %s %s", modname, fnl_path))
+      return true
     elseif true then
       local __1_auto = _19_
-      return FAIL(string.format("Can require module %s %s", modname, fnl_path))
+      FAIL(string.format("Can require module %s %s", modname, fnl_path))
+      return false
     else
       return nil
     end
