+++ test/utils.fnl
@@ -21,7 +21,7 @@
   (print "\n")
   (os.exit results.fails))
 
-(vim.opt.runtimepath:prepend "/hotpot")
+(vim.opt.runtimepath:prepend (vim.loop.cwd))
 (require :hotpot)
 
 {: write-file : read-file
